Refactor swagger schema
