+++ src/app.ts
@@ -19,8 +19,9 @@ import * as routes from "./constants/urls";
 import { errorHandler } from "./middlewares/errorHandler";
 import { deserializeUser } from "./middlewares/deserializeUser";
 
-import swaggerJSDoc from "swagger-jsdoc";
 import swaggerUi from "swagger-ui-express";
+
+import { swaggerSpec } from "./docs/docConfig/swaggerConfig";
 
 declare global {
   namespace Express {
@@ -31,21 +32,6 @@ declare global {
 }
 
 const app = express();
-
-const options = {
-  definition: {
-    openapi: "3.0.1",
-    info: {
-      title: "Vaccine Management REST API Documentation",
-      version: "1.0.0",
-    },
-    schemes: ["http", "https"],
-    servers: [{ url: process.env.BASE_URL }],
-  },
-  apis: [`${__dirname}/routes/*.ts`, "./dist/routes/*.js"],
-};
-
-const swaggerSpec = swaggerJSDoc(options);
 
 dotenv.config();
