+++ src/docs/docSchemas/tokenSchemaDoc.ts
@@ -1,0 +1,20 @@
+/**
+ * @swagger
+ * components:
+ *   schemas:
+ *     RefreshToken:
+ *       type: object
+ *       properties:
+ *        refreshToken:
+ *           type: string
+ *           description: The user's refresh token.
+ *           example: zxczxczxSDSDsdxczczxdasdDSAFDSA
+ *
+ *     AccessToken:
+ *       type: object
+ *       properties:
+ *        accessToken:
+ *           type: string
+ *           description: The user's access token.
+ *           example: zxczxczxSDSDsdxczczxdasdDSAFDSA
+ */
