+++ src/routes/userRoute.ts
@@ -8,67 +8,6 @@ const router = Router();
 
 /**
  * @swagger
- * components:
- *   schemas:
- *     User:
- *       type: object
- *       properties:
- *         name:
- *           type: string
- *           description: The user's name.
- *           example: Admin
- *         gender:
- *           type: string
- *           description: The user's gender.
- *           example: male
- *         dateOfBirth:
- *           type: date
- *           description: The user's date of birth.
- *           example: 1997/02/10
- *         email:
- *           type: email
- *           description: The user's email.
- *           example: admin@test.com
- *         address:
- *           type: string
- *           description: The user's address.
- *           example: USA
- *
- *     LoginPayload:
- *       type: object
- *       properties:
- *         email:
- *           type: email
- *           description: The user's email.
- *           example: admin@test.com
- *         password:
- *           type: string
- *           description: The user's password.
- *           example: password
- *
- *     UserLoggedIn:
- *       allOf:
- *         - type: object
- *           properties:
- *             id:
- *               type: integer
- *               description: The user ID.
- *               example: 0
- *         - $ref: '#/components/schemas/User'
- *
- *     UserSigningUp:
- *       allOf:
- *         - type: object
- *           properties:
- *             password:
- *               type: string
- *               description: The user's password.
- *               example: password
- *         - $ref: '#/components/schemas/User'
- */
-
-/**
- * @swagger
  * /signup:
  *   post:
  *     summary: Create a new user.
@@ -89,7 +28,7 @@ const router = Router();
  *               type: object
  *               properties:
  *                 user:
- *                   $ref: '#/components/schemas/UserLoggedIn'
+ *                   $ref: '#/components/schemas/UserLoggingIn'
  *                 message:
  *                   type: string
  *                   example: User created successfully
@@ -119,12 +58,6 @@ router.post(routes.SIGN_UP, userController.signUp);
  *               properties:
  *                 user:
  *                   $ref: '#/components/schemas/UserLoggedIn'
- *                 accessToken:
- *                   type: string
- *                   example: zxcxcadESDFXCVXCVZzxczxcszxcsxzx
- *                 refreshToken:
- *                   type: string
- *                   example: axzcsadwsdfSDFSCVxcvxcvxcvesdfes
  *                 message:
  *                   type: string
  *                   example: User logged in successfully
